Updates Gradle, Kotlin, and dependencies
Updates Gradle wrapper to version 8.4 and Kotlin to version 1.9.10.

Migrates localization to use `flutter_gen` package, generating localizations in `lib/l10n`, and updates relevant imports throughout the application.

This change ensures that the project is using the latest versions of the build tools and dependencies, and correctly configures localization.